preserve flag + fail checks

+++ VGM2WAV.py
@@ -8,6 +8,7 @@ parser = argparse.ArgumentParser()
 parser.add_argument("files", nargs="*", help="Files to process")
 parser.add_argument("-f", "--formats", nargs="+", default=default_formats, choices=all_formats,
                     help="List of formats to check for. Separated by spaces.")
+parser.add_argument("--preserve", action="store_true", help="Preserve original files after conversion.")
 
 
 def get_files_from_path(path, formats):
@@ -33,15 +34,24 @@ def collect_all_files(paths, formats):
         files.extend(get_files_from_path(path, formats))
     return files
 
-def executeFiles(exe_path, binka_files):
-    for file in binka_files:
-        command = [str(exe_path), str(file), '-o', str(file.with_suffix(".wav"))]
+def executeFiles(exe_path, files, PRESERVE: bool = False):
+    for file in files:
+        output_file = file.with_suffix(".wav")
+        command = [str(exe_path), str(file), '-o', str(output_file)]
         print(subprocess.list2cmdline(command))
 
         try:
             subprocess.run(command, check=True)
-            file.unlink()
-            print(f"Converted: {file}")
+            if PRESERVE:
+                continue # skip delete if flag is set
+
+            if output_file.exists() and output_file.stat().st_size > 0:
+                file.unlink()
+                print(f"Converted: {file}")
+            else:
+                raise RuntimeError("Process completed with no errors thrown by vgmstream, however output file is either not written or empty.\n" \
+                                f"Original will be preserved in case further action is necessary: \n {file}")
+            
         except subprocess.CalledProcessError as e:
             print(f"Conversion failed: {file}")
             print(e)
@@ -64,7 +74,7 @@ if __name__ == "__main__":
     files = collect_all_files(args.files, args.formats)
 
     if files:
-        executeFiles(exe_path, files)
+        executeFiles(exe_path, files, args.preserve)
 
     else:
         print("No supported files found.")
